Prepare CellBasedScheme for LinearHet and NonLinear
