+++ demos/highorder/main.cpp
@@ -14,9 +14,9 @@ namespace fs = std::filesystem;
 template <class Field,
           // scheme config
           std::size_t neighbourhood_width = 2,
-          class cfg                       = samurai::StarStencilFV<Field::dim, Field::size, neighbourhood_width>,
-          class bdry_cfg                  = samurai::BoundaryConfigFV<neighbourhood_width>>
-class HighOrderDiffusion : public samurai::CellBasedScheme<cfg, bdry_cfg, Field>
+          class cfg      = samurai::StarStencilFV<samurai::SchemeType::LinearHomogeneous, Field::size, neighbourhood_width, Field>,
+          class bdry_cfg = samurai::BoundaryConfigFV<neighbourhood_width>>
+class HighOrderDiffusion : public samurai::CellBasedScheme<cfg, bdry_cfg>
 {
   public:
