+++ include/samurai/schemes/fv/cell_based_scheme.hpp
@@ -1,140 +1,44 @@
 #pragma once
+#include "../explicit_scheme.hpp"
 #include "FV_scheme.hpp"
+#include "flux_definition.hpp"
 
 namespace samurai
 {
+    enum class SchemeType
+    {
+        NonLinear,
+        LinearHeterogeneous,
+        LinearHomogeneous
+    };
+
     /**
-     * Useful sizes to define the sparsity pattern of the matrix and perform the preallocation.
+     * @class CellBasedScheme
      */
-    template <std::size_t output_field_size_,
-              std::size_t neighbourhood_width_,
-              std::size_t scheme_stencil_size_,
-              std::size_t center_index_,
-              std::size_t contiguous_indices_start_ = 0,
-              std::size_t contiguous_indices_size_  = 0>
-    struct CellBasedAssemblyConfig
+    template <class cfg, class bdry_cfg, class check = void>
+    class CellBasedScheme
     {
-        static constexpr std::size_t output_field_size        = output_field_size_;
-        static constexpr std::size_t neighbourhood_width      = neighbourhood_width_;
-        static constexpr std::size_t scheme_stencil_size      = scheme_stencil_size_;
-        static constexpr std::size_t center_index             = center_index_;
-        static constexpr std::size_t contiguous_indices_start = contiguous_indices_start_;
-        static constexpr std::size_t contiguous_indices_size  = contiguous_indices_size_;
     };
 
-    template <std::size_t dim, std::size_t output_field_size, std::size_t neighbourhood_width = 1>
-    using StarStencilFV = CellBasedAssemblyConfig<output_field_size,
-                                                  neighbourhood_width,
-                                                  // ---- Stencil size
-                                                  // Cell-centered Finite Volume scheme:
-                                                  // center + 'neighbourhood_width' neighbours in each Cartesian direction (2*dim
-                                                  // directions) --> 1+2=3 in 1D
-                                                  //                 1+4=5 in 2D
-                                                  1 + 2 * dim * neighbourhood_width,
-                                                  // ---- Index of the stencil center
-                                                  // (as defined in star_stencil())
-                                                  neighbourhood_width,
-                                                  // ---- Start index and size of contiguous cell indices
-                                                  // (as defined in star_stencil())
-                                                  0,
-                                                  1 + 2 * neighbourhood_width>;
-
-    template <std::size_t output_field_size>
-    using OneCellStencilFV = CellBasedAssemblyConfig<output_field_size,
-                                                     0,  // Neighbourhood width
-                                                     1,  // Stencil size
-                                                     0>; // Index of the stencil center
-
-    template <std::size_t output_field_size>
-    using EmptyStencilFV = CellBasedAssemblyConfig<output_field_size,
-                                                   0,  // Neighbourhood width
-                                                   0,  // Stencil size
-                                                   0>; // Index of the stencil center
-
-    template <class cfg, class bdry_cfg, class Field>
-    class CellBasedScheme : public FVScheme<Field, cfg::output_field_size, bdry_cfg>
+    template <class cfg>
+    auto make_cell_based_scheme(/*const FluxDefinition<cfg>& flux_definition*/)
     {
-      protected:
+        using bdry_cfg = BoundaryConfigFV<cfg::stencil_size / 2>;
 
-        using base_class = FVScheme<Field, cfg::output_field_size, bdry_cfg>;
-        using base_class::dim;
-        using base_class::field_size;
+        return CellBasedScheme<cfg, bdry_cfg>(); // flux_definition);
+    }
 
-      public:
-
-        using cfg_t                                    = cfg;
-        using bdry_cfg_t                               = bdry_cfg;
-        using field_t                                  = Field;
-        using field_value_type                         = typename Field::value_type; // double
-        static constexpr std::size_t output_field_size = cfg::output_field_size;
-        using local_matrix_t                           = typename detail::LocalMatrix<field_value_type,
-                                                            output_field_size,
-                                                            field_size>::Type; // 'double' if field_size = 1, 'xtensor' representing a
-                                                                                                         // matrix otherwise
-        using scheme_stencil_t      = Stencil<cfg::scheme_stencil_size, dim>;
-        using get_coefficients_func = std::function<std::array<local_matrix_t, cfg::scheme_stencil_size>(double)>;
-
-      private:
-
-        scheme_stencil_t m_stencil;
-        get_coefficients_func m_get_coefficients;
-
-      public:
-
-        explicit CellBasedScheme()
-        {
-        }
-
-        auto& stencil() const
-        {
-            return m_stencil;
-        }
-
-        auto& stencil()
-        {
-            return m_stencil;
-        }
-
-        void set_stencil(const scheme_stencil_t& stencil)
-        {
-            m_stencil = stencil;
-        }
-
-        void set_stencil(scheme_stencil_t&& stencil)
-        {
-            m_stencil = stencil;
-        }
-
-        get_coefficients_func& coefficients_func() const
-        {
-            return m_get_coefficients;
-        }
-
-        get_coefficients_func& coefficients_func()
-        {
-            return m_get_coefficients;
-        }
-
-        void set_coefficients_func(get_coefficients_func get_coefficients)
-        {
-            m_get_coefficients = get_coefficients;
-        }
-
-        auto coefficients(double h) const
-        {
-            return m_get_coefficients(h);
-        }
-    };
-
+    /**
+     * is_CellBasedScheme
+     */
     template <class Scheme, typename = void>
     struct is_CellBasedScheme : std::false_type
     {
     };
 
     template <class Scheme>
-    struct is_CellBasedScheme<
-        Scheme,
-        std::enable_if_t<std::is_base_of_v<CellBasedScheme<typename Scheme::cfg_t, typename Scheme::bdry_cfg_t, typename Scheme::field_t>, Scheme>>>
+    struct is_CellBasedScheme<Scheme,
+                              std::enable_if_t<std::is_base_of_v<CellBasedScheme<typename Scheme::cfg_t, typename Scheme::bdry_cfg_t>, Scheme>>>
         : std::true_type
     {
     };
