+++ include/samurai/schemes/fv/operators/identity_FV.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "../cell_based_scheme.hpp"
+#include "../cell_based_scheme__lin_hom.hpp"
 
 namespace samurai
 {
@@ -8,10 +8,10 @@ namespace samurai
     {
         static constexpr std::size_t dim = Field::dim;
 
-        using cfg      = OneCellStencilFV<Field::size>;
+        using cfg      = OneCellStencilFV<SchemeType::LinearHomogeneous, Field::size, Field>;
         using bdry_cfg = BoundaryConfigFV<1>;
 
-        CellBasedScheme<cfg, bdry_cfg, Field> identity;
+        CellBasedScheme<cfg, bdry_cfg> identity;
 
         using local_matrix_t = typename decltype(identity)::local_matrix_t;
